# api/openapi.yaml
openapi: 3.1.0
info:
  title: Дом в MAX — API коллективных заявок
  version: 1.0.0
  description: |
    API мини-приложения и бота: жители сообщают о проблемах с общим имуществом дома,
    присоединяются к существующим заявкам, УК меняет статусы.

    Вход: POST /auth/max (initData мини-приложения) или POST /auth/demo (демо-роли для проверки).
    Дальше заголовок `Authorization: Bearer <token>`, сессия живёт 12 часов.

    Ошибки всегда в формате `{"error": {"code": "...", "message": "..."}}`.
servers:
  - url: /api/v1
...
paths:
...
  /auth/max:
    post:
      tags: [auth]
      summary: Вход по initData мини-приложения MAX
      description: Сервер проверяет подпись HMAC-SHA256 и свежесть auth_date (не старше 1 часа). Новый пользователь MAX становится жителем.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [init_data]
              properties:
                init_data: { type: string, description: Строка window.WebApp.initData }
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "401": { $ref: "#/components/responses/Error" }
  /auth/demo:
    post:
      tags: [auth]
      summary: Демо-вход для проверки без клиента MAX
      description: Работает только при DEMO_AUTH_ENABLED=true.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [resident, resident_2, uk_operator, district]
                  description: resident и resident_2 — два соседа из одного дома, uk_operator — сотрудник УК
      responses:
        "200": { $ref: "#/components/responses/Session" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /classify:
    post:
      tags: [issues]
      summary: Подсказка категории по тексту жителя
      description: |
        Если подключена self-hosted модель (OLLAMA_URL), она выбирает код из закрытого справочника;
        при ошибке, таймауте или неизвестном коде работают ключевые слова. Житель подтверждает категорию сам,
        ответственный и срок задают правила. Все поля null, если категорию не узнали.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1, maxLength: 1000, example: Не горит свет на пятом этаже }
      responses:
        "200":
          description: Подсказка
          content:
            application/json:
              schema:
                type: object
                required: [category, title, source]
                properties:
                  category: { type: [string, "null"], example: lighting }
                  title: { type: [string, "null"], example: Свет в подъезде }
                  source: { type: [string, "null"], enum: [llm, rules, null] }
        "401": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /me:
    get:
      tags: [me]
      summary: Текущий пользователь
      responses:
        "200": { $ref: "#/components/responses/User" }
        "401": { $ref: "#/components/responses/Error" }
    delete:
      tags: [me]
      summary: Удалить аккаунт
      description: |
        Имя, телефон, согласие, дом и связь с аккаунтом MAX стираются, фото пользователя удаляются;
        заявки остаются без данных автора. Вернувшийся житель получает новый аккаунт без старых заявок.
        Демо-пользователь при следующем демо-входе возвращается в исходное состояние.
      responses:
        "204": { description: Аккаунт удалён }
        "401": { $ref: "#/components/responses/Error" }
  /me/phone:
    post:
      tags: [me]
      summary: Оставить телефон для мастера
      description: |
        Номер из WebApp.requestContact() с подписью клиента MAX: hash = HMAC_SHA256("authDate=…\nphone=…\nuserId=…",
        токен бота), телефон без «+». Подпись не старше часа. Номер — это и согласие показывать его УК по заявкам,
        где житель участник. Только для входа через MAX: демо-пользователю 403.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, auth_date, hash]
              properties:
                phone: { type: string, example: "+79991234567" }
                auth_date: { type: string, description: Как вернул клиент MAX (секунды или миллисекунды) }
                hash: { type: string, description: hex }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
    delete:
      tags: [me]
      summary: Не показывать телефон
      responses:
        "200": { $ref: "#/components/responses/User" }
        "401": { $ref: "#/components/responses/Error" }
...
  /me/consent:
    post:
      tags: [me]
      summary: Дать согласие на обработку персональных данных
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                version: { type: string, example: v1 }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "422": { $ref: "#/components/responses/Error" }
  /me/house:
    post:
      tags: [me]
      summary: Привязать жителя к дому
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [house_id]
              properties:
                house_id: { type: string, example: h-17k2 }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "404": { $ref: "#/components/responses/Error" }
  /houses:
    get:
      tags: [houses]
      summary: Поиск дома по адресу
      parameters:
        - { name: query, in: query, required: true, schema: { type: string, minLength: 2 }, example: 17к2 }
      responses:
        "200":
          description: До 20 домов
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/House" }
        "422": { $ref: "#/components/responses/Error" }
  /houses/nearest:
    get:
      tags: [houses]
      summary: Ближайшие дома по координатам
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number }, example: 55.6124 }
        - { name: lon, in: query, required: true, schema: { type: number }, example: 37.7462 }
      responses:
        "200":
          description: До 5 домов, ближайший первым
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/House" }
        "422": { $ref: "#/components/responses/Error" }
  /geo/reverse:
    get:
      tags: [houses]
      summary: Адрес точки по данным OpenStreetMap
      description: |
        Подсказка к «Найти дома рядом». Геокодер совместим с Nominatim (GEOCODER_URL).
        Если геокодер выключен, не нашёл адрес или не ответил за 3 с, ответ 200 без полей.
        Рядом с адресом нужно показывать attribution: этого требует лицензия ODbL.
      parameters:
        - { name: lat, in: query, required: true, schema: { type: number }, example: 55.6128 }
        - { name: lon, in: query, required: true, schema: { type: number }, example: 37.7199 }
      responses:
        "200":
          description: Адрес или пустой объект
          content:
            application/json:
              schema:
                type: object
                properties:
                  address: { type: string, example: "Ореховый бульвар, 15" }
                  attribution: { type: string, example: "© участники OpenStreetMap" }
        "401": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /houses/{id}:
    get:
      tags: [houses]
      summary: Карточка дома с УК, подъездами и объектами
      parameters:
        - { name: id, in: path, required: true, schema: { type: string }, example: h-17k2 }
      responses:
        "200":
          description: Дом
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HouseDetails" }
        "404": { $ref: "#/components/responses/Error" }
  /houses/{id}/issues:
    get:
      tags: [issues]
      summary: Заявки дома, новые первыми
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/IssueList" }
  /objects/{code}:
    get:
      tags: [houses]
      summary: Вход по QR-коду на объекте
      parameters:
        - { name: code, in: path, required: true, schema: { type: string }, example: h-17k2-e2-lift }
      responses:
        "200":
          description: Объект и его дом
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { $ref: "#/components/schemas/AssetObject" }
                  house: { $ref: "#/components/schemas/House" }
        "404": { $ref: "#/components/responses/Error" }
  /issues/similar:
    get:
      tags: [issues]
      summary: Открытые похожие заявки за 14 дней (защита от дублей)
      parameters:
        - { name: house_id, in: query, required: true, schema: { type: string } }
        - { name: category, in: query, required: true, schema: { type: string } }
        - { name: object_id, in: query, required: false, schema: { type: string } }
      responses:
        "200": { $ref: "#/components/responses/IssueList" }
        "422": { $ref: "#/components/responses/Error" }
  /issues:
    post:
      tags: [issues]
      summary: Сообщить о проблеме
      description: |
        Ответственный — УК дома, срок — из справочника категорий в рабочих днях (по Москве).
        Если указан object_id, категория и заголовок по умолчанию берутся из объекта.
        Нужно согласие на обработку ПДн.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [house_id]
              properties:
                house_id: { type: string, example: h-17k2 }
                object_id: { type: string, example: h-17k2-e2-lift }
                category: { type: string, example: lift }
                title: { type: string }
                description: { type: string, example: Кабина не приходит на вызов }
      responses:
        "201":
          description: Заявка создана
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /issues/{id}:
    get:
      tags: [issues]
      summary: Карточка заявки с адресом, ответственным и основанием срока
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Заявка
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
        "404": { $ref: "#/components/responses/Error" }
  /issues/{id}/timeline:
    get:
      tags: [issues]
      summary: Хронология заявки для карточки
      description: Кто именно действовал, не раскрывается.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: События по времени
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    kind: { type: string, enum: [created, joined, status_changed, overdue, confirmed, reopened] }
                    status: { $ref: "#/components/schemas/Status" }
                    comment: { type: string }
                    at: { type: string, format: date-time }
        "404": { $ref: "#/components/responses/Error" }
  /issues/{id}/join:
    post:
      tags: [issues]
      summary: «Это и у меня» — присоединиться к заявке
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Заявка с обновлённым числом участников
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
        "403": { $ref: "#/components/responses/Error" }
        "409":
          description: already_joined или issue_closed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /issues/{id}/status:
    post:
      tags: [uk]
      summary: Сменить статус (оператор ответственной УК)
      description: |
        Переходы: sent → accepted | in_progress | rejected; accepted → in_progress | done | rejected;
        in_progress → done | rejected. Для rejected комментарий обязателен.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: "#/components/schemas/Status" }
                comment: { type: string, example: Мастер приедет завтра до 12:00 }
      responses:
        "200":
          description: Заявка с новым статусом
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /issues/{id}/confirm:
    post:
      tags: [issues]
      summary: Подтвердить, что починили
      description: |
        Житель, участник выполненной заявки, в течение 7 дней после отметки «выполнено» (answer_until), один ответ
        на каждое «выполнено». Уведомлений не рассылает. 403 не житель (сотрудник УК, район) или не участник;
        409 not_done, window_closed, already_answered. Подавать заявки и присоединяться тоже может только житель.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Заявка с числом подтверждений
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /issues/{id}/reopen:
    post:
      tags: [issues]
      summary: Не починили — вернуть заявку в работу
      description: |
        Те же условия, что у confirm, плюс обязательный комментарий (422 comment_required).
        Заявка переходит в in_progress, срок считается заново по справочнику, просрочка сбрасывается,
        reopened_at фиксирует возврат. Соседи получают сообщение в чате с ботом.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [comment]
              properties:
                comment: { type: string, maxLength: 1000, example: На третьем этаже снова темно }
      responses:
        "200":
          description: Заявка снова в работе
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Issue" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /issues/{id}/appeal:
    post:
      tags: [issues]
      summary: Ссылка на черновик обращения в жилинспекцию (PDF)
      description: |
        Только участнику открытой заявки, у которой истёк срок ответа. Ссылка подписана и действует 10 минут;
        по ней PDF скачивается без заголовка авторизации (в MAX — через WebApp.downloadFile).
        409 not_overdue — срок ещё не истёк; 409 issue_closed — заявка закрыта; 403 — не участник.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Ссылка на PDF
          content:
            application/json:
              schema:
                type: object
                required: [url, expires_at, file_name]
                properties:
                  url: { type: string, example: /api/v1/appeal/aWQ.c2ln }
                  expires_at: { type: string, format: date-time }
                  file_name: { type: string, example: obrashchenie-142.pdf }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /issues/{id}/photos:
    get:
      tags: [issues]
      summary: Фото заявки
      description: Только участникам заявки и сотрудникам ответственной УК.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Фото в порядке загрузки
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Photo" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [issues]
      summary: Приложить фото к открытой заявке
      description: |
        Участник или сотрудник ответственной УК; житель — с действующим согласием (иначе 403 consent_required).
        До 3 файлов за запрос в поле photo, JPEG или PNG до 5 МБ и 40 Мп, не больше 6 фото на заявку.
        Пачка сохраняется целиком или не сохраняется: один плохой файл отклоняет все.
        Сервер уменьшает снимок до 1600 px по длинной стороне, поворачивает по EXIF
        и перекодирует в JPEG без EXIF: геометка и данные камеры не сохраняются.
        413 photo_too_large, 415 unsupported_media, 409 too_many_photos или issue_closed.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [photo]
              properties:
                photo:
                  type: array
                  maxItems: 3
                  items: { type: string, format: binary }
      responses:
        "201":
          description: Загруженные фото
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Photo" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /photos/{id}:
    get:
      tags: [issues]
      summary: Файл фото (JPEG)
      description: |
        Только с токеном сессии; мини-приложение загружает файл через fetch, а не <img src>.
        Cache-Control: private, no-store — после удаления снимок не остаётся в WebView.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Фото
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [issues]
      summary: Убрать своё фото
      description: Только тот, кто приложил фото (в списке у него mine = true); даже УК не убирает чужие.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Фото удалено }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /appeal/{token}:
    get:
      tags: [issues]
      summary: PDF-обращение по подписанной ссылке
      description: |
        Ссылка живёт 10 минут. Условия проверяются заново: заявку за это время закрыли (409 issue_closed)
        или аккаунт владельца ссылки удалён (403).
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Черновик обращения
          content:
            application/pdf:
              schema: { type: string, format: binary }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /uk/issues:
    get:
      tags: [uk]
      summary: Очередь заявок УК
      description: Открытые по сроку (просроченные первыми), затем закрытые. Только для uk_operator.
      responses:
        "200": { $ref: "#/components/responses/IssueList" }
        "403": { $ref: "#/components/responses/Error" }
  /uk/houses:
    get:
      tags: [uk]
      summary: Дома своей УК
      description: Только для uk_operator. Нужны для печати наклеек с QR-кодами объектов (объекты — в GET /houses/{id}).
      responses:
        "200":
          description: Дома по адресу
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/House" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
  /district/overdue:
    get:
      tags: [district]
      summary: Просроченные заявки района
      description: Открытые заявки домов района с истёкшим сроком, самые давние первыми. Только для роли district.
      responses:
        "200": { $ref: "#/components/responses/IssueList" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
...
security:
  - bearer: []
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  responses:
    Session:
      description: Сессия
      content:
        application/json:
          schema:
            type: object
            properties:
              token: { type: string }
              expires_at: { type: string, format: date-time }
              user: { $ref: "#/components/schemas/User" }
              start_param: { type: string, description: Параметр запуска мини-приложения, например o_<qr_code> }
    User:
      description: Пользователь
      content:
        application/json:
          schema: { $ref: "#/components/schemas/User" }
    IssueList:
      description: Список заявок
      content:
        application/json:
          schema:
            type: array
            items: { $ref: "#/components/schemas/Issue" }
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum: [unauthorized, consent_required, forbidden, not_found, already_joined, issue_closed,
                     invalid_transition, not_overdue, too_many_photos, not_done, window_closed, already_answered,
                     photo_too_large, unsupported_media, reason_required, comment_required, invalid_contact,
                     invalid_input, http_error, internal]
            message: { type: string, description: Текст для показа жителю }
...
    Organization:
      type: object
      properties:
        id: { type: string }
        type: { type: string, enum: [uk, tsj, rso, regoperator, capital_repair_fund, gji] }
        name: { type: string }
        phone_office: { type: string }
        phone_dispatcher: { type: string }
        phone_emergency: { type: string }
        schedule: { type: string }
    House:
      type: object
      properties:
        id: { type: string }
        address: { type: string }
        district: { type: string }
        year_built: { type: integer }
        floors: { type: integer }
        entrances_count: { type: integer }
        organization_id: { type: string }
        source: { type: string, description: "model — модельная запись, в интерфейсе помечается" }
    AssetObject:
      type: object
      properties:
        id: { type: string }
        house_id: { type: string }
        entrance_id: { type: string }
        category: { type: string }
        label: { type: string, example: "подъезд 2, пассажирский лифт" }
        qr_code: { type: string }
    HouseDetails:
      allOf:
        - $ref: "#/components/schemas/House"
        - type: object
          properties:
            organization: { $ref: "#/components/schemas/Organization" }
            entrances:
              type: array
              items:
                type: object
                properties:
                  id: { type: string }
                  number: { type: integer }
            objects:
              type: array
              items: { $ref: "#/components/schemas/AssetObject" }
    Status:
      type: string
      enum: [sent, accepted, in_progress, done, rejected]
    Photo:
      type: object
      required: [id, url, width, height, mine, created_at]
      properties:
        id: { type: string, format: uuid }
        url: { type: string, example: /api/v1/photos/0190a000-0000-7000-8000-000000000001 }
        width: { type: integer, example: 1600 }
        height: { type: integer, example: 1200 }
        mine: { type: boolean, description: "Фото приложил текущий пользователь и может его убрать; кто приложил чужое, не раскрывается" }
        created_at: { type: string, format: date-time }
...